Updates for new range and zip functions
In Python 2, range and zip return lists, but now they return iterators.
In a few cases we actually wanted the lists to be lists, so we now have
to explicitly construct lists from the iterators in those cases.  In
other cases we wanted iterators, which in Python 2 had different names,
so this commit updates those old names for Python 3.