Fix schedcat.util.math.topsum
The behavior of map() changed in Python 3, causing our topsum() function
to break.  This commit fixes it by simplifying the implementation using
heapq.nlargest() rather than sorting the whole array.  This commit also
changes topsum() to not call map(), taking instead only an iterable and
the number of values to add.  In the one place in the codebase where the
map() feature was used, a generator expression is now used instead.

This commit additionally removes two duplicated tests of lcm() from the
test method for topsum(), and adds another topsum() test.