Use __lt__ instead of __le__ for heapq
Python 2 allowed heaps to work with __le__ if __lt__ was undefined, but
Python 3 no longer does this.  This broke the PrioObj in
schedcat.util.iter, which was based on __le__ rather than __lt__.  This
commit changes the PrioObj to define __lt__ instead of __le__, fixing it
for Python 3 and not breaking any existing consumer code (it's really
just a semantic change, and in fact some code defined the "le" function
using <).